+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## [1.0.4] - 24 April 2020
+    - Bug Solved(_searchInput returns a future inside setstate)
+
 ## [1.0.3] - 20 Feb 2020
-    - Mapbox Search dependency updated to latest 
+    - Mapbox Search dependency updated to latest
 
 ## [1.0.1] - 30 Jan 2020